split up payout.go file
